Extract Scanner and InputScanner into external classes
In this commit we extracted the Scanner and InputScanner classes into
their own file, reducing the code under the `CSV::Parser` class.

Deliberately no other change was made, to ensure we do not break
compatibility.
